backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
structlog